+++ meda-mailbox/src/App.tsx
@@ -1,9 +1,9 @@
 // ======================================================================
 import { PagePathEnum } from "./app/common/enums/page-paths.enum";
 
-import { 
+import {
   Route, Link,
-  BrowserRouter, Routes, 
+  BrowserRouter, Routes,
 } from "react-router-dom";
 import HomePage from "./app/pages/HomePage";
 import InboxPage from "./app/pages/InboxPage";
@@ -16,12 +16,8 @@ function App() {
   return (
     <BrowserRouter>
       <Routes>
-        {/* 
-        // TODO: switch inbox and home; home displays the `view message` button.
-        // TODO: inbox displays the messages
-        */}
-        <Route path='/' element={<InboxPage />} />
-        <Route path={`/${PagePathEnum.INBOX}`} element={<HomePage />}>
+        <Route path='/' element={<HomePage />} />
+        <Route path={`/${PagePathEnum.INBOX}`} element={<InboxPage />}>
           <Route path={`${PagePathEnum.SPLIT_VIEW}`} element={<SplitViewPage />} />
           <Route path={`${PagePathEnum.PANEL_VIEW}`} element={<SingleViewPage />} />
           <Route path={`${PagePathEnum.INBOX_MESSAGE_VIEW}`} element={<MessageViewPage />} />
@@ -29,12 +25,12 @@ function App() {
         <Route
           path="*"
           element={
-            <main style={{ padding: "1rem" }}>
-              <p>There's nothing here!</p>
+            <>
+              <p>Guard pending to be implemented!</p>
               <Link to="/">
                 <button> Go to Inbox </button>
               </Link>
-            </main>
+            </>
           }
         />
       </Routes>
