+++ meda-mailbox/src/app/pages/InboxPage.tsx
@@ -1,43 +1,19 @@
 // =============================================================
-import '../common/common.css';
+import { Outlet } from 'react-router-dom';
+import { FeedbackFeat } from '../features/FeedbackFeat';
 
-import { Link } from 'react-router-dom';
-import { ButtonComp } from '../components/ButtonComp';
-import { PagePathEnum } from '../common/enums/page-paths.enum';
-import { MessageCounterFeat } from '../features/MessageCounterFeat';
-import { MUI_WARNING_BUTTON } from '../common/constants/button.constant';
-
-import Box from "@mui/material/Box/Box";
-import Grid from "@mui/material/Grid/Grid";
+import Container from '@mui/material/Container';
+import NavbarFeat from '../features/NavbarFeat';
 // =============================================================
 
-const boxStyle = {
-  width: '100wh',
-  height: '100vh',
-  backgroundColor: 'primary.main',
-  '&:hover': {
-    backgroundColor: 'primary.dark',
-  },
-}
-
-function Inbox() {
-  return (
-    <Grid container spacing={2} className='text-align-center'>
-      <Grid item xs={12} sm={12} md={12} lg={12} xl={12}>
-        <Box className='full-vector-height-width' sx={boxStyle}>
-          <MessageCounterFeat />
-          <Link to={PagePathEnum.INBOX_SPLIT_VIEW}>
-            <ButtonComp
-              text="View Messages"
-              mui={MUI_WARNING_BUTTON}
-            />
-          </Link>
-        </Box>
-      </Grid>
-    </Grid>
-  )
-}
+const Home = () => (
+  <Container maxWidth="md">
+    <NavbarFeat />
+    <Outlet />
+    <FeedbackFeat />
+  </Container>
+)
 
 export default function InboxPage() {
-  return <Inbox />
+  return <Home />
 }
